refactor: combine commands and settings
This will combine the commands and there settings into a single file to
make working more easier with them.
