+++ src/IL2CareerToolset/Commands/Cli/ManuelDatabaseCommand.cs
@@ -1,5 +1,4 @@
 ﻿using IL2CareerModel.Services;
-using IL2CareerToolset.Commands.Cli.Settings;
 using IL2CareerToolset.Services;
 using Microsoft.Extensions.Logging;
 using Spectre.Console;
@@ -7,6 +6,15 @@ using Spectre.Console.Cli;
 using System.ComponentModel;
 using System.Diagnostics.CodeAnalysis;
 namespace IL2CareerToolset.Commands.Cli;
+
+internal class ManuelDatabaseCommandSettings : CommandSettings
+{
+    [CommandArgument(0, "[GameRootFolder]")]
+    [Description("Path to the root folder of the game")]
+    public string? GameRootFolder { get; set; }
+
+    public bool IsInteractiv => GameRootFolder is null;
+}
 
 [Description("Command to set the game path by yourself")]
 internal class ManuelDatabaseCommand : Command<ManuelDatabaseCommandSettings>
