+++ src/IL2CareerToolset/Commands/Cli/Save/DeleteBackupsCommand.cs
@@ -1,13 +1,22 @@
 ﻿using IL2CareerModel.Models.Database;
 using IL2CareerModel.Services;
-using IL2CareerToolset.Commands.Cli.Settings;
-using Microsoft.Extensions.Logging;
 using Spectre.Console;
 using Spectre.Console.Cli;
 using System.ComponentModel;
 using System.Diagnostics.CodeAnalysis;
 
 namespace IL2CareerToolset.Commands.Cli.Save;
+
+internal class DeleteBackupsCommandSettings : CommandSettings
+{
+    [CommandOption("-a|--all")]
+    [Description("Delete all backups")]
+    public bool All { get; set; }
+
+    [CommandArgument(0, "[BACKUP_GUID]")]
+    [Description("The backup guid to delete")]
+    public string? BackupGuid { get; set; }
+}
 
 [Description("Delete a single or multiple backups")]
 internal class DeleteBackupsCommand : Command<DeleteBackupsCommandSettings>
