+++ src/IL2CareerToolset/Commands/Cli/SetLogLevelCommand.cs
@@ -1,4 +1,4 @@
-﻿using IL2CareerToolset.Commands.Cli.Settings;
+﻿using IL2CareerToolset.Enums;
 using IL2CareerToolset.Services;
 using Microsoft.Extensions.Logging;
 using Spectre.Console;
@@ -7,6 +7,15 @@ using System.ComponentModel;
 using System.Diagnostics.CodeAnalysis;
 
 namespace IL2CareerToolset.Commands.Cli;
+
+internal class SetLogLevelCommandSettings : CommandSettings
+{
+    private const string POSSIBLE_LOG_LEVELS = "Trace, Debug, Information, Warning, Error, Critical";
+
+    [CommandArgument(0, "<LOG_LEVEL>")]
+    [Description($"The log level to use for the program. Possible values are {POSSIBLE_LOG_LEVELS}")]
+    public LogSeverity LogLevel { get; set; }
+}
 
 [Description("Set the log level for the application. If you set it on Information every log for information and above will be shown")]
 internal class SetLogLevelCommand : Command<SetLogLevelCommandSettings>
